In progress SeaORM -> SQLX & Sea Query migration

+++ src/entities/figure.rs
@@ -1,10 +1,13 @@
 use serde::{Serialize};
+use sea_query::enum_def;
 
-#[derive(Serialize, Debug)]
+#[allow(dead_code)]
+#[enum_def(suffix = "Def")]
+#[derive(Serialize, Debug, sqlx::FromRow)]
 pub struct Figure {
-    pub id: i32,
+    pub id: i64,
     pub title: String,
     pub width: Option<i32>,
     pub height: Option<i32>,
-    pub profile_id: i32,
+    pub profile_id: i64,
 }
